requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
asgiref==3.8.1
    # via django
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
    # via environs
dj-email-url==1.0.6
    # via environs
django==5.0.4
    # via
    #   crispy-bootstrap5
    #   dj-database-url
    #   django-allauth
    #   django-crispy-forms
    #   django-debug-toolbar
django-allauth==0.63.6
django-cache-url==3.4.5
    # via environs
django-crispy-forms==2.2
    # via crispy-bootstrap5
django-debug-toolbar==4.3.0
environs==11.0.0
marshmallow==3.21.3
    # via environs
packaging==24.1
    # via marshmallow
python-dotenv==1.0.1
    # via environs
shortuuid==1.0.13
sqlparse==0.5.0
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.12.2
    # via dj-database-url
